// cgocore-ui/src/app/features/models/bookings/views/full/BookingFullView.ts
import { BookingForm } from 'src/app/features/models/bookings/forms/BookingForm';
/**
 * auto generated full veiw bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class BookingFullView {

id: number;
form: BookingForm;
canUpdate: boolean;
canDelete: boolean;

/* Specific Code Start */
/* Specific Code End */
}